requirements.txt
appwrite==15.0.0
beautifulsoup4==4.14.3
playwright==1.58.0
playwright-stealth==2.0.1
python-dotenv==1.2.1
requests==2.32.5
lxml==6.0.2
tavily-python==0.7.21
